backend/requirements-core.txt
fastapi==0.135.1
PyJWT==2.12.1
python-dotenv==1.2.2
pydantic==2.12.5
uvicorn==0.27.0
redis==7.3.0
sse-starlette==3.0.3
httpx==0.28.1
pyyaml==6.0.3
prometheus-client==0.24.1
pywebpush==2.3.0
gunicorn==25.1.0
sqlalchemy[asyncio]==2.0.48
pgvector==0.4.2
asyncpg==0.31.0
alembic[tz]==1.18.4
bcrypt==5.0.0
python-multipart==0.0.22
webauthn==2.7.1
aiomqtt==2.5.1
opentelemetry-api==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-exporter-otlp-proto-http==1.33.0
